refactor(transcript): trim redundant test machinery
The client has no transparent page replay path, while the Host already keeps query results replayable until overlay acknowledgement. Remove the synthetic identical-page tolerance and its per-page serialization cost.\n\nShare the duplicated transcript reader fixture and drop tests that only asserted private SQLite layout or mock callback counts, while retaining coverage for observable paging, integrity, recovery, and resource-lifecycle contracts.
